Complete make over (#23)
* Complete make over

Use mongodb instead of sql since it better fits the purpose.
Add timesheet because it is part of the domain language
Update major version to 1
Trash postman collection. Will use swagger.
Remove h2 from readme.
Use list in dto because it is easier to process via JSON.
Use month as id for timesheet since it is the natural key.
Minor improvements.